some renaming and moving of utilities. some pr changes. added back in threads for forward passes - did some testing and saw a significant speed up.
