--- sup3r/utilities/test_utils.py
@@ -1,383 +1,0 @@
-"""Utilities used across multiple test files"""
-import os
-import numpy as np
-import xarray as xr
-from PIL import Image
-
-from sup3r.postprocessing.file_handling import OutputHandlerH5
-from sup3r.utilities.utilities import pd_date_range
-
-
-def make_fake_nc_files(td, input_file, n_files):
-    """Make dummy nc files with increasing times
-
-    Parameters
-    ----------
-    input_file : str
-        File to use as template for all dummy files
-    n_files : int
-        Number of dummy files to create
-
-    Returns
-    -------
-    fake_files : list
-        List of dummy files
-    """
-    fake_dates = [f'2014-10-01_{str(i).zfill(2)}_00_00'
-                  for i in range(n_files)]
-    fake_times = [f'2014-10-01 {str(i).zfill(2)}:00:00'
-                  for i in range(n_files)]
-    fake_files = [os.path.join(td, f'input_{date}') for date in fake_dates]
-    for i in range(n_files):
-        input_dset = xr.open_dataset(input_file)
-        with xr.Dataset(input_dset) as dset:
-            dset['Times'][:] = np.array([fake_times[i].encode('ASCII')],
-                                        dtype='|S19')
-            dset['XTIME'][:] = i
-            dset.to_netcdf(fake_files[i])
-    return fake_files
-
-
-def make_fake_h5_chunks(td):
-    """Make fake h5 chunked output files for a 5x spatial 2x temporal
-    multi-node forward pass output.
-
-    Parameters
-    ----------
-    td : tempfile.TemporaryDirectory
-        Test TemporaryDirectory
-
-    Returns
-    -------
-    out_files : list
-        List of filepaths to chunked files.
-    data : ndarray
-        (spatial_1, spatial_2, temporal, features)
-        High resolution forward pass output
-    ws_true : ndarray
-        Windspeed between 0 and 20 in shape (spatial_1, spatial_2, temporal, 1)
-    wd_true : ndarray
-        Windir between 0 and 360 in shape (spatial_1, spatial_2, temporal, 1)
-    features : list
-        List of feature names corresponding to the last dimension of data
-        ['windspeed_100m', 'winddirection_100m']
-    slices_lr : list
-        List of low res temporal slices corresponding to the out_files list
-    slices_hr : list
-        List of high res temporal slices corresponding to the out_files list
-    low_res_lat_lon : ndarray
-        Array of lat/lon for input data. (spatial_1, spatial_2, 2)
-        Last dimension has ordering (lat, lon)
-    low_res_times : list
-        List of np.datetime64 objects for coarse data.
-    """
-
-    features = ['windspeed_100m', 'winddirection_100m']
-    model_meta_data = {'foo': 'bar'}
-    shape = (50, 50, 96, 1)
-    ws_true = np.random.uniform(0, 20, shape)
-    wd_true = np.random.uniform(0, 360, shape)
-    data = np.concatenate((ws_true, wd_true), axis=3)
-    lat = np.linspace(90, 0, 10)
-    lon = np.linspace(-180, 0, 10)
-    lon, lat = np.meshgrid(lon, lat)
-    low_res_lat_lon = np.dstack((lat, lon))
-
-    low_res_times = pd_date_range('20220101', '20220103', freq='3600s',
-                                  inclusive='left')
-
-    slices_lr = [slice(0, 24), slice(24, None)]
-    slices_hr = [slice(0, 48), slice(48, None)]
-
-    out_files = [os.path.join(td, 'fp_out_0.h5'),
-                 os.path.join(td, 'fp_out_1.h5')]
-
-    for i, (slice_lr, slice_hr) in enumerate(zip(slices_lr, slices_hr)):
-        OutputHandlerH5.write_output(data[:, :, slice_hr, :], features,
-                                     low_res_lat_lon,
-                                     low_res_times[slice_lr],
-                                     out_files[i],
-                                     model_meta_data,
-                                     max_workers=1)
-
-    out = (out_files, data, ws_true, wd_true, features, slices_lr, slices_hr,
-           low_res_lat_lon, low_res_times)
-
-    return out
-
-
-def make_fake_cs_ratio_files(td, low_res_times, low_res_lat_lon, gan_meta):
-    """Make a set of dummy clearsky ratio files that match the GAN fwp outputs
-
-    Parameters
-    ----------
-    td : tempfile.TemporaryDirectory
-        Test TemporaryDirectory
-    low_res_times :
-        List of times for low res input data. If there is only a single low
-        res timestep, it is assumed the data is daily.
-    low_res_lat_lon
-        Array of lat/lon for input data.
-        (spatial_1, spatial_2, 2)
-        Last dimension has ordering (lat, lon)
-    gan_meta : dict
-        Meta data for model to write to file.
-
-    Returns
-    -------
-    fps : list
-        List of clearsky ratio .h5 chunked files.
-    fp_pattern : str
-        Glob pattern*string to find fps
-    """
-    fps = []
-    chunk_dir = os.path.join(td, 'chunks/')
-    fp_pattern = os.path.join(chunk_dir, 'sup3r_chunk_*.h5')
-    os.makedirs(chunk_dir)
-
-    for idt, timestamp in enumerate(low_res_times):
-        fn = ('sup3r_chunk_{}_{}.h5'
-              .format(str(idt).zfill(6), str(0).zfill(6)))
-        out_file = os.path.join(chunk_dir, fn)
-        fps.append(out_file)
-
-        cs_ratio = np.random.uniform(0, 1, (20, 20, 1, 1))
-        cs_ratio = np.repeat(cs_ratio, 24, axis=2)
-
-        OutputHandlerH5.write_output(cs_ratio, ['clearsky_ratio'],
-                                     low_res_lat_lon,
-                                     [timestamp],
-                                     out_file, max_workers=1,
-                                     meta_data=gan_meta)
-    return fps, fp_pattern
-
-
-def tke_spectrum(u, v):
-    """Longitudinal Turbulent Kinetic Energy Spectrum. Gives the portion of
-    kinetic energy associated with each longitudinal wavenumber.
-
-    Parameters
-    ----------
-    u: ndarray
-        (lat, lon)
-        U component of wind
-    v : ndarray
-        (lat, lon)
-        V component of wind
-
-    Returns
-    -------
-    ndarray
-        1D array of amplitudes corresponding to the portion of total energy
-        with a given longitudinal wavenumber
-    """
-    v_k = np.fft.fftn(v)
-    u_k = np.fft.fftn(u)
-    E_k = np.abs(v_k)**2 + np.abs(u_k)**2
-    E_k = np.mean(E_k, axis=0)
-    E_k *= np.arange(E_k.shape[0])**2
-    n_steps = E_k.shape[0] // 2
-    E_k_a = E_k[:n_steps]
-    E_k_b = E_k[-n_steps:][::-1]
-    E_k = E_k_a + E_k_b
-    return E_k
-
-
-def tke_series(u, v):
-    """Longitudinal Turbulent Kinetic Energy Spectrum time series. Gives the
-    mean tke spectrum over time.
-
-    Parameters
-    ----------
-    u: ndarray
-        (lat, lon, time)
-        U component of wind
-    v : ndarray
-        (lat, lon, time)
-        V component of wind
-
-    Returns
-    -------
-    ndarray
-        1D array of mean tke amplitudes over time
-    """
-
-    return [np.mean(tke_spectrum(u[..., t], v[..., t]))
-            for t in range(u.shape[-1])]
-
-
-def velocity_gradient_dist(u, bins=50, range=None):
-    """Returns the longitudinal velocity gradient distribution.
-
-    Parameters
-    ----------
-    u: ndarray
-        Longitudinal velocity component
-        (lat, lon, temporal)
-    bins : int
-        Number of bins for the velocity gradient pdf.
-    range : tuple | None
-        Optional min/max range for the velocity gradient pdf.
-
-    Returns
-    -------
-    ndarray
-        du/dx at bin centers
-    ndarray
-        Normalized delta_u / delta_x value counts
-    """
-    diffs = np.diff(u, axis=1).flatten()
-    diffs = diffs[(np.abs(diffs) < 7)]
-    diffs = diffs / np.sqrt(np.mean(diffs**2))
-    counts, edges = np.histogram(diffs, bins=bins, range=range)
-    centers = edges[:-1] + (np.diff(edges) / 2)
-    counts = counts.astype(float) / counts.max()
-    return centers, counts
-
-
-def vorticity_dist(u, v, bins=50, range=None):
-    """Returns the vorticity distribution.
-
-    Parameters
-    ----------
-    u: ndarray
-        Longitudinal velocity component
-        (lat, lon, temporal)
-    v : ndarray
-        Latitudinal velocity component
-        (lat, lon, temporal)
-    bins : int
-        Number of bins for the vorticity pdf.
-    range : tuple | None
-        Optional min/max range for the vorticity pdf.
-
-    Returns
-    -------
-    ndarray
-        vorticity values at bin centers
-    ndarray
-        Normalized vorticity value counts
-    """
-    dudy = np.diff(u, axis=0, append=np.mean(u)).flatten()
-    dvdx = np.diff(v, axis=1, append=np.mean(v)).flatten()
-    diffs = dudy - dvdx
-    diffs = diffs[(np.abs(diffs) < 14)]
-    diffs = diffs / np.sqrt(np.mean(diffs**2))
-    counts, edges = np.histogram(diffs, bins=bins, range=range)
-    centers = edges[:-1] + (np.diff(edges) / 2)
-    counts = counts.astype(float) / counts.max()
-    return centers, counts
-
-
-def ramp_rate_dist(u, v, bins=50, range=None):
-    """Returns the ramp rate distribution.
-
-    Parameters
-    ----------
-    u: ndarray
-        Longitudinal velocity component
-        (lat, lon, temporal)
-    v : ndarray
-        Latitudinal velocity component
-        (lat, lon, temporal)
-    bins : int
-        Number of bins for the ramp rate pdf.
-    range : tuple | None
-        Optional min/max range for the ramp rate pdf.
-
-    Returns
-    -------
-    ndarray
-        dws/dt values at bin centers
-    ndarray
-        Normalized delta_ws / delta_t value counts
-    """
-    diffs = np.diff(np.sqrt(u**2 + v**2), axis=-1).flatten()
-    diffs = diffs[(np.abs(diffs) < 10)]
-    diffs = diffs / np.sqrt(np.mean(diffs**2))
-    counts, edges = np.histogram(diffs, bins=bins, range=range)
-    centers = edges[:-1] + (np.diff(edges) / 2)
-    counts = counts.astype(float) / counts.max()
-    return centers, counts
-
-
-def spatial_interp(low, s_enhance):
-    """Spatial bilinear interpolation for low resolution field. Used to provide
-    baseline for comparison with gan output
-
-    Parameters
-    ----------
-    low : ndarray
-        Low resolution field to interpolate.
-        (spatial_1, spatial_2, temporal)
-    s_enhance : int
-        Factor by which to enhance the spatial domain
-
-    Returns
-    -------
-    ndarray
-        Spatially interpolated low resolution output
-    """
-
-    high = np.zeros((low.shape[0] * s_enhance, low.shape[1] * s_enhance,
-                     low.shape[-1]))
-    for t in range(low.shape[-1]):
-        im = Image.fromarray(low[..., t])
-        im = im.resize((low[..., t].shape[1] * s_enhance,
-                        low[..., t].shape[0] * s_enhance),
-                       resample=Image.Resampling.BILINEAR)
-        high[..., t] = np.array(im)
-    return high
-
-
-def temporal_interp(low, t_enhance):
-    """Temporal bilinear interpolation for low resolution field. Used to
-    provide baseline for comparison with gan output
-
-    Parameters
-    ----------
-    low : ndarray
-        Low resolution field to interpolate.
-        (spatial_1, spatial_2, temporal)
-    t_enhance : int
-        Factor by which to enhance the temporal domain
-
-    Returns
-    -------
-    ndarray
-        Temporally interpolated low resolution output
-    """
-
-    high = np.zeros((low.shape[0], low.shape[1], low.shape[-1] * t_enhance))
-    for t in range(high.shape[-1]):
-        t0 = t // t_enhance
-        t1 = t0 + 1
-        alpha = (t / t_enhance - t0) / (t1 - t0)
-        if t1 == low.shape[-1]:
-            tmp = low[..., t0]
-        else:
-            tmp = (1 - alpha) * low[..., t0] + alpha * low[..., t1]
-        high[..., t] = tmp
-    return high
-
-
-def st_interp(low, s_enhance, t_enhance):
-    """Spatiotemporal bilinear interpolation for low resolution field. Used to
-    provide baseline for comparison with gan output
-
-    Parameters
-    ----------
-    low : ndarray
-        Low resolution field to interpolate.
-        (spatial_1, spatial_2, temporal)
-    s_enhance : int
-        Factor by which to enhance the spatial domain
-    t_enhance : int
-        Factor by which to enhance the temporal domain
-
-    Returns
-    -------
-    ndarray
-        Spatiotemporally interpolated low resolution output
-    """
-    return temporal_interp(spatial_interp(low, s_enhance), t_enhance)
